ref(integration-tests): Restructure test projects to follow Amplify Gen 2 conventions
Split schema/resource.ts into auth/resource.ts and data/resource.ts
following the standard Amplify Gen 2 directory layout. Backends now
import auth and data from their conventional locations.
