+++ packages/test-infra/amplify/backend.ts
@@ -19,15 +19,12 @@ backend.auth.resources.cfnResources.cfnUserPool.addPropertyOverride(
     true,
 );
 
-backend.auth.resources.cfnResources.cfnUserPoolClient.addPropertyOverride(
-    "ExplicitAuthFlows",
-    [
-        "ALLOW_CUSTOM_AUTH",
-        "ALLOW_USER_SRP_AUTH",
-        "ALLOW_REFRESH_TOKEN_AUTH",
-        "ALLOW_USER_PASSWORD_AUTH",
-    ],
-);
+backend.auth.resources.cfnResources.cfnUserPoolClient.addPropertyOverride("ExplicitAuthFlows", [
+    "ALLOW_CUSTOM_AUTH",
+    "ALLOW_USER_SRP_AUTH",
+    "ALLOW_REFRESH_TOKEN_AUTH",
+    "ALLOW_USER_PASSWORD_AUTH",
+]);
 
 const testUsersSecretName = "overtone-test-users";
